prop: inv_ult_bvprop: Use interval computation.
Determine inverse value with respect to propagator domain via interval
computation. That is, first we propagate the target value from t to x
via s by utilizing propagator domains, and then select a value (or
detect a conflict) with respect to the resulting domain based on the
interval of possible values for x.