sat: Add support for Kissat Extras
This adds support for using the Kissat Extras fork which adds supports
incremental solving.

Using Kissat and Kissat Extras at the same time is not possible as they
define the same symbols. Thus this reuses blzakissat.{c,h} and the used
Kissat version is selected by running either of the two setup scripts.
Support for incremental solving is detected at compile time via a define
set by Kissat Extras.

As there hasn't been a stable release of Kissat Extras yet, the setup
script pulls the latest version from github and does output a warning
that this is a development snapshot.

Signed-off-by: Jannis Harder <me@jix.one>