sat: Update to Kissat 2.0.1
There has been a new Kissat release (sc2021-sweep) from the SAT
Competition 2021. Additionally there have been two minor releases since:

  * 2.0.0: Fixing a build issue by inlining kissat_signal_name. That
    issues reproduces on my setup.
  * 2.0.1: Removing an unconditional runtime assertion ("coverage
    goal").

As the minor releases are not tagged in the repository, this uses the
full commit hash in the setup script. The changes since the last tagged
sc2021-sweep release are small, so it is not too hard to verify that
there are no changes in solver behavior.

Signed-off-by: Jannis Harder <me@jix.one>