bounds: and: Prototype for considering signed and unsigned bounds.
This implements considering both signed and unsigned bounds when
determining is_invertible. It introduces a new kind of domain generator
(BitVectorDomainDualGenerator), which may be used as a more general
version of BitVectorDomainSignedGenerator in the future. It further
prototypes how to determine the actual bounds given both signed and
unsigned bounds in BitVectorAnd::is_invertible(). This will be
refactored and generalized for all node kinds in follow-up commits.