prop: Refactor path selection to ensure deterministic execution with logging.
Previously, when logging was enabled, this triggered additional calls to
Node::is_essential(), which triggered additional calls to the RNG and
resulted in non-deterministic execution between runs with and without
logging. This refactors path selection to return the set of inputs
determined to be essential, which allows to log this information while
preserving determinism.