smt2 parser: Allow lists of attributes in annotations.
Previously, not the full SMT-LIB annotation syntax was supported, it was
only possible to annotate a term with a single attribute. Now, a list of
attributes (as defined in the standard as (! <term> <attribute>+) is
supported. Note that still only attribute :named is supported, anything
other than :named is ignored.