interpolation: Refactor construction of interpolant.
Previously, the interpolant was still constructed in CNF form. This is
redundant as AND gates are already reconstructed from the CNF given by
CaDiCraig. However, we have already reconstructed the Tseitin
representing the root of the AIG representation of the interpolant.
Note that constructing the interpolant in CNF does not only introduce
overhead for the construction but also doesn't entirely simplify to the
root AND node we have reconstructed by rewriting.