meson: Reconfigure CryptoMiniSat subproject to not use cmake.
Using cmake if CryptoMiniSat is not found in the system and built from
source breaks configuration between system and from source. This is due
to inconsistent setup of the CryptoMiniSat build system on install. We
now switch to defining our own meson build configuration for
CryptoMiniSat and do not use cmake to build the subproject.