interpolation: Refactor labeling of SAT variables.
Previously, we labeled all A and C variables as shared and lemmas as A.
Now we follow the proper definition of interpolation: uninterpreted
symbol that occur in the interpolant may not be local to A or C, they
must be shared. As a consequence, we currently cannot compute
interpolants when arrays and uninterpreted functions are involved
(lemmas may introduce 'mixed' terms, i.e., terms with occurrences of
A-local and C-local symbols).