sat: Make configuring terminator for Kissat, CMS fatal.
We now throw an exception when configuring a terminator and the
configured SAT solver does not have terminator support (CryptoMinisat
and Kissat). Previously, we allowed configuration of a terminator and
only issued a warning on check_sat(), which is deceiving for users as we
will terminate on the Bitwuzla layer via the configured terminator, but
the SAT solver will ignore the configured termination conditions.