meson: Build GMP by default for static builds.
GMP is a dependency for MPFR and needs to be available during the
configure phase of MPFR. As a result, we can not build GMP as a fallback
project anymore during Bitwuzla's build phase.

For static builds, we now build static versions for GMP and MPFR by
default and do not rely on the versions provided by the system.
This also avoids linking issues encountered on macOS with static
libraries installed in non-standard locations via homebrew.