interpolants: Fix incremental handling.
When solving incrementally, then all clauses that encode assumptions
remain in the SAT solver. Thus, it can happen that "inactive" clauses
and literals are included in the proof core (it is not minimal).
However, inactive clauses and SAT variables are not labeled, and must
not be considered when generating the interpolant. This fixes interpolant
generation for this case.