meson: don't require Git as a compile-time dependency
Python's subprocess.run throws a FileNotFound exception if the executed
command was not found. Similarly, meson's run_command will error if the
command was not found (even when run with `check=False`.

This commit works around these issue by (a) handling the FileNotFound
exception and (b) invoking Git via `sh(1)` so that a non-zero exit
status is returned if Git isn't installed (which is then handled by
`check=False`).

This was discovered while updating Guix's bitwuzla package.

Previously: #114