interpolants: Sanitize interpolant via rewrite.
At the moment, it may happen that the bit-level interpolant is correct
and only refers to bits of GLOBAL symbols, but when backmapping, we pull
in local symbols. A symptom of this is when, e.g., when the interpolant
contains extracts on original terms. This fix is more of a bandaid,
albeit a good one. It pushes extracts down as far as possible via
rewriting, by introducing an interpolant-specific rewriter that is
applied on the interpolant after standard rewriting. This does not fix
all possible cases that may occur, though.