interpolants: Only map back consts.
To avoid that we pull local symbols into the interpolant when mapping
back the SAT interpolant to the word-level, we now only map back AIG consts
to slices of word-level consts and build the node-level interpolant from
the AIG on top. This results in interpolants that are more bit-level
than word-level, with a potentially (very likely) larger size. In the
future, we hope to be able to extract more word-level structure.