fp: Fix issue with missing validity constraints with incremental preprop.
In the incremental case, it was possible that validity constraints for
FP/RM consts and leafs were registered with the bit-blast solver but
never bit-blasted. This only happened if a BV engine other than the
bit-blast engine solved an incremental query where consts/leafs were
word-blasted but the corresponding assertions were popped. This fix
ensures that such missing validity constraints are readded.