interpolants: Revert to clearing assertions when not producing interpolants.
When we don't produce interpolants, we revert to our default handling
of assertions in the bitblast solver. Previously, since on pop we
destroy and reinitialize the SAT solver when producing interpolants,
we had to keep d_assertions around (to readd the assertions to the SAT
solver). When not producing interpolants, this is redundant and only
introduces overhead.